Add guards to migration

+++ db/migrate/20231201202449_remove_grading_deadline_from_assessments.rb
@@ -1,11 +1,13 @@
 class RemoveGradingDeadlineFromAssessments < ActiveRecord::Migration[6.0]
   def up
-    remove_column :assessments, :grading_deadline
+    remove_column :assessments, :grading_deadline if column_exists?(:assessments, :grading_deadline)
   end
   def down
-    add_column :assessments, :grading_deadline, :datetime, null: false
-    Assessment.find_each do |asmt|
-      asmt.update_attribute(:grading_deadline, asmt.end_at)
+    unless column_exists?(:assessments, :grading_deadline)
+      add_column :assessments, :grading_deadline, :datetime, null: false
+      Assessment.find_each do |asmt|
+        asmt.update_attribute(:grading_deadline, asmt.end_at)
+      end
     end
   end
 end
